rename Parser module to ParserD
The top level module "Parser" will be using both ParserK and ParserD.
